better_tomorrow/requirements.txt
Django==3.1
djangorestframework==3.12.2
requests==2.28.1
